requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
numpy==2.0.0
opencv-python-headless==4.10.0.84
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
